backend-python/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.3
ecdsa==0.19.2
fastapi==0.136.1
h11==0.16.0
idna==3.15
pyasn1==0.6.3
pydantic==2.13.4
pydantic_core==2.46.4
python-jose==3.5.0
python-multipart==0.0.28
rsa==4.9.1
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
